+++ app/src/main/java/com/example/weather/models/WeatherDetails.kt
@@ -41,6 +41,6 @@ data class Sys(
     val type: Int,
     val id: Int,
     val country: String,
-    val sunrise: Long,
-    val sunset: Long
+    val sunrise: Int,
+    val sunset: Int
 )
